# std.mem.indexOfScalarPos: ~20% faster