This patch simply cuts in half the number of checks for matches in the wider SIMD code path, so the measured impact is on larger files.

The ~20% claim comes from the following test:
```zig
pub fn main() !void {
    var buf: [4096]u8 = undefined;
    var file = try std.fs.cwd().openFile("test", .{});
    defer file.close();

    const slice = buf[0..try file.readAll(buf[0..])];

    var indx: usize = 0;
    for (0..4096) |s| {
        for (0..255) |v| {
            indx = indx + (std.mem.indexOfScalarPos(u8, slice, s, @truncate(v)) orelse 0);
        }
    }
    std.debug.print("{}\n", .{indx});
}
```

The test file is 4KiB and is composed of random alphanumerical characters 

```
Benchmark 1 (242 runs): ./new
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          20.0ms ±  949us    19.8ms … 33.9ms         13 ( 5%)        0%
  peak_rss            782KB ±  588       774KB …  782KB          2 ( 1%)        0%
  cpu_cycles         89.1M  ± 4.20M     88.3M  …  151M          28 (12%)        0%
  instructions        322M  ± 3.23       322M  …  322M          15 ( 6%)        0%
  cache_references    840   ± 1.12K      374   … 16.7K           9 ( 4%)        0%
  cache_misses       43.1   ±  104         0   … 1.42K          10 ( 4%)        0%
  branch_misses       273K  ± 1.41K      258K  …  276K           4 ( 2%)        0%
Benchmark 2 (197 runs): ./main
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          24.7ms ±  945us    24.5ms … 36.1ms          4 ( 2%)        💩+ 23.4% ±  0.9%
  peak_rss            807KB ± 2.92KB     766KB …  807KB          1 ( 1%)        💩+  3.1% ±  0.0%
  cpu_cycles          110M  ± 4.17M      109M  …  160M          13 ( 7%)        💩+ 23.5% ±  0.9%
  instructions        395M  ± 5.29       395M  …  395M          12 ( 6%)        💩+ 22.7% ±  0.0%
  cache_references    926   ± 1.01K      443   … 13.5K           5 ( 3%)          + 10.2% ± 24.0%
  cache_misses       59.0   ±  132         4   … 1.42K          11 ( 6%)          + 36.9% ± 51.2%
  branch_misses       242K  ±  960       237K  …  245K           5 ( 3%)        ⚡- 11.2% ±  0.1%
```
The processor is an i7-7700K CPU @ 4.20GHz