@nektro 2025-08-10
can you edit the benchmark so that `./main` is fist/baseline?