@andrewrk 2025-08-13
A nice property of the implementation here is that it obeys pointer provenance rules. It's technically undefined behavior to assume you can read past a memory allocation, even if you stay within a page boundary.

I'll keep working on the language spec and compiler implementation to try and resolve this problem, but in the meantime, it would be good to not assume you can read past a memory allocation.