@GiuseppeCesarano 2025-08-13
Simply removing the `=`  sign in the while should resolve the thing, the benchmarks are within 1% of the previous results.