@GiuseppeCesarano 2025-08-20
I've spent some time reading how llvm treats BE vectors and investigating the issue, i can confirm it's not a bug in `std.simd.firstTrue`, it's output is wrong in the test but the function itself should not account for BE vectors which is a backend concern. 

Another thing which I've noticed is that one of the failing tests `priority_queue.test.siftUp in remove` fails under `aarch64_be-linux-musl` debug, but it passes if compiled in release mode, while also passing for `aarch64_be-linux-gnu` in debug mode. 

I've checked the llvm-ir for those targets, the release mode has a slightly different ir for the `std.mem.indexOfScalarPos` that handles 4 x u32 function while the two debug modes are identical, same for the `std.simd.firstTrue` 4 x i1 funciton, it has the same ir.

Still, only the musl debug version shows the behavior that counts lanes from right.
 
 Could the problem be how the ir gets lowered to asm under the musl-debug target?